Move data processing to 'processData' in 'dataUtils.js'
