+++ app/src/App.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import Plot from 'react-plotly.js';
+import { processData } from './dataUtils';
 
 const App = () => {
   // Store plot data in state.
@@ -15,16 +16,8 @@ const App = () => {
         // 'data.json' should be populated from a run of sim.py
         const response = await fetch('data.json');
         const data = await response.json();
-        const updatedPlotData = {};
-
-        data.forEach(([t0, t1, frame]) => {
-          for (let [agentId, { x, y }] of Object.entries(frame)) {
-            updatedPlotData[agentId] = updatedPlotData[agentId] || { x: [], y: [] };
-            updatedPlotData[agentId].x.push(x);
-            updatedPlotData[agentId].y.push(y);
-          }
-        });
-
+        // Process data into format for plotting
+        const updatedPlotData = processData(data);
         setPlotData(Object.values(updatedPlotData));
         console.log('plotData:', Object.values(updatedPlotData));
       } catch (error) {
@@ -40,7 +33,7 @@ const App = () => {
       style={{ position: 'fixed', width: '100%', height: '100%', left: 0, top: 0 }}
       data={plotData}
       layout={{
-        title: 'Visualization',
+        title: 'Orbit Simulation',
         yaxis: { scaleanchor: 'x' },
         autosize: true,
       }}
